pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "coverage>=7.10.6",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
]
[tool.setuptools.packages.find]
where = ["src"]
